requirements_mac.txt
setuptools >= 21.0.0
gunicorn==22.0.0
flask-marshmallow
marshmallow-sqlalchemy
flask-cors
connexion[swagger-ui]
connexion[flask]
connexion[uvicorn]
flask-testing
numpy<2.0
rasterio==1.4.3
Pillow==11.1.0
geopandas==1.0.1
transformers==4.49.0
torch>=2.2.2
huggingface_hub==0.29.3
pytest
ipdb
PyYAML>=5.1
matplotlib
ultralytics
segment-geospatial